+++ example-app/src/main/java/com/example/app/MainActivity.java
@@ -21,6 +21,7 @@ import com.globalpayments.library.terminals.moby.MobyDevice;
 import com.globalpayments.library.utilities.PermissionHelper;
 import com.globalpayments.library.utilities.PermissionHelper.PermissionsCallBack;
 import com.tsys.payments.library.gateway.enums.GatewayType;
+import timber.log.Timber;
 
 public class MainActivity extends BaseActivity implements View.OnClickListener, PermissionsCallBack {
 
@@ -66,8 +67,6 @@ public class MainActivity extends BaseActivity implements View.OnClickListener, 
     private Button disconnect;
     private Switch safSwitch;
     private Switch surchargeSwitch;
-    private Switch surchargePretaxSwitch;
-    private EditText surchargeCustomPercent;
     private int startCounter = 0;
 
     public static boolean isShowAbout() {
@@ -143,9 +142,6 @@ public class MainActivity extends BaseActivity implements View.OnClickListener, 
         safSwitch.setOnClickListener(this);
         surchargeSwitch = findViewById(R.id.surcharge_switch);
         surchargeSwitch.setOnClickListener(this);
-        surchargePretaxSwitch = findViewById(R.id.surcharge_pretax_switch);
-        surchargePretaxSwitch.setOnClickListener(this);
-        surchargeCustomPercent = findViewById(R.id.surcharge_custom_edittext);
         about = findViewById(R.id.about_button);
         disconnect = findViewById(R.id.disconnect_button);
         disconnect.setOnClickListener(this);
@@ -266,6 +262,11 @@ public class MainActivity extends BaseActivity implements View.OnClickListener, 
                 if (mobyDevice == null) {
                     mobyDevice =
                             new MobyDevice(getApplicationContext(), getConnectionConfig(ConnectionMode.BLUETOOTH));
+                    // Logs handled via lambda expression
+                    mobyDevice.setLogCallback(logDataJson -> {
+                        // Handle the log data here
+                        Timber.tag(TAG).d("MobyDeviceLog : %s", logDataJson.toString());
+                    });
                 } else {
                     mobyDevice.setConnectionConfig(getConnectionConfig(ConnectionMode.BLUETOOTH));
                 }
@@ -284,6 +285,8 @@ public class MainActivity extends BaseActivity implements View.OnClickListener, 
             try {
                 if (mobyDevice == null) {
                     mobyDevice = new MobyDevice(getApplicationContext(), getConnectionConfig(ConnectionMode.USB));
+                    // Logs handled via Handler class
+                    mobyDevice.setLogCallback(new ExampleGPLibraryLogHandler());
                 } else {
                     mobyDevice.setConnectionConfig(getConnectionConfig(ConnectionMode.USB));
                 }
@@ -323,6 +326,8 @@ public class MainActivity extends BaseActivity implements View.OnClickListener, 
         } else if (view.getId() == R.id.disconnect_button) {
             if (mobyDevice != null && mobyDevice.isConnected()) {
                 mobyDevice.disconnect();
+                // Remove callback on device disconnect
+                mobyDevice.removeLogCallback();
                 mobyDevice = null;
                 startCounter = 1;
                 showHideButtonDisplay(false, DeviceType.MOBY);
@@ -346,12 +351,6 @@ public class MainActivity extends BaseActivity implements View.OnClickListener, 
             } else {
                 Toast.makeText(this, "Surcharge is disabled", Toast.LENGTH_LONG).show();
             }
-        } else if (view.getId() == R.id.surcharge_pretax_switch) {
-            if (surchargePretaxSwitch.isChecked()) {
-                Toast.makeText(this, "Surcharge Pre-Tax is enabled", Toast.LENGTH_LONG).show();
-            } else {
-                Toast.makeText(this, "Surcharge Pre-Tax is disabled", Toast.LENGTH_LONG).show();
-            }
         }
     }
 
@@ -374,25 +373,9 @@ public class MainActivity extends BaseActivity implements View.OnClickListener, 
         connectionConfig.setSafEnabled(safSwitch.isChecked());
         connectionConfig.setSafExpirationInDays(5);
         connectionConfig.setSurchargeEnabled(surchargeSwitch.isChecked());
-        connectionConfig.setSurchargePreTax(surchargePretaxSwitch.isChecked());
-        if (!surchargeCustomPercent.getText().toString().isEmpty()) {
-            connectionConfig.setSurchargePercent(getCustomSurchargePercent());
-        }
         connectionConfig.setEnvironment(environmentSwitch.isChecked() ? Environment.PRODUCTION : Environment.TEST);
         connectionConfig.setGateway(gatewaySwitch.isChecked() ? GatewayType.TRANSIT : GatewayType.PORTICO);
         return connectionConfig;
-    }
-
-    private float getCustomSurchargePercent() {
-        String customPercent = surchargeCustomPercent.getText().toString();
-        float floatValue;
-        try {
-            floatValue = Float.parseFloat(customPercent);
-        } catch (NumberFormatException e) {
-            e.printStackTrace();
-            return 0;
-        }
-        return floatValue;
     }
 
     /**